add Pricing email

+++ code/routes/web.php
@@ -23,42 +23,48 @@ use App\Http\Controllers\AERequestTableController;
 
 //Admin Routes
 Route::prefix('admin')->group(function () {
+
+    //Login And Register Routes Admin
     Route::get('/login',[AdminLoginController::class,'index'])->name('admin.index');
     Route::post('/login/owner',[AdminLoginController::class,'checklogin'])->name('admin.login');
     Route::get('/dashboard',[AdminLoginController::class,'dashboard'])->name('admin.dashboard')->middleware('admin');
     Route::post('/logout',[AdminLoginController::class,'adminlogout'])->name('admin.logout')->middleware('admin');
 
+    //AE  Request Form Admin
     Route::get('/ae-requestform',[AERequestController::class,'index'])->name('AEAdmin')->middleware('admin');
     Route::get('/ae-requestform/create',[AERequestController::class,'create'])->middleware('admin');
     Route::post('/ae-requestform/store',[AERequestController::class,'store'])->middleware('admin');
     Route::get('/ae-requestform/{id}',[AERequestController::class,'show'])->middleware('admin');
     Route::put('/ae-requestform/{id}',[AERequestController::class,'update'])->middleware('admin');
-    Route::put('/ae-requestform/status/{id}',[AERequestController::class,'ae_change_status']);
+    Route::put('/ae-requestform/status/{id}',[AERequestController::class,'ae_change_status'])->middleware('admin');
 
+    //AE Table Routes Admin
     Route::get('/requesttable',[AERequestTableController::class,'index'])->name('AETable')->middleware('admin');
 
+    //Pricing Routes Admin
     Route::get('/pricing-form',[PricingRateController::class,'index'])->name('pricing')->middleware('admin');
     Route::get('/pricing-form/create',[PricingRateController::class,'create'])->middleware('admin');
     Route::post('/pricing-form',[PricingRateController::class,'store'])->middleware('admin');
 
+    //Billing Routes Admin
     Route::get('/billing-form',[BillingController::class,'index'])->name('billing')->middleware('admin');
     Route::get('/billing-form/create',[BillingController::class,'create'])->middleware('admin');
 
-
+    //Role Routes Admin
     Route::get('/role',[URolesController::class,'index'])->name('role')->middleware('admin');
-    Route::post('/role', [URolesController::class,'store']);
-    Route::get('/role/create', [URolesController::class,'create']);
-    Route::get('/role/{id}', [URolesController::class,'show']);
-    Route::put('/role/{id}', [URolesController::class,'update']);
-    Route::delete('/role/{id}', [URolesController::class,'destroy']);
+    Route::post('/role', [URolesController::class,'store'])->middleware('admin');
+    Route::get('/role/create', [URolesController::class,'create'])->middleware('admin');
+    Route::get('/role/{id}', [URolesController::class,'show'])->middleware('admin');
+    Route::put('/role/{id}', [URolesController::class,'update'])->middleware('admin');
+    Route::delete('/role/{id}', [URolesController::class,'destroy'])->middleware('admin');
 
-    //user_managment->user
+    //User Routes Admin
     Route::get('/user', [UUserController::class,'index'])->name('user')->middleware('admin');
-    Route::post('/user', [UUserController::class,'store']);
-    Route::get('/user/create', [UUserController::class,'create']);
-    Route::get('/user/{id}', [UUserController::class,'show']);
-    Route::put('/user/{id}', [UUserController::class,'update']);
-    Route::delete('/user/{id}', [UUserController::class,'destroy']);
+    Route::post('/user', [UUserController::class,'store'])->middleware('admin');
+    Route::get('/user/create', [UUserController::class,'create'])->middleware('admin');
+    Route::get('/user/{id}', [UUserController::class,'show'])->middleware('admin');
+    Route::put('/user/{id}', [UUserController::class,'update'])->middleware('admin');
+    Route::delete('/user/{id}', [UUserController::class,'destroy'])->middleware('admin');
 });
 
 Route::get('/', function () {
